requirements.txt
autogen
pyautogen
chainlit
psycopg2-binary
openai
typing-extensions